Move some utility functions from setup one level up
